refactor: extract rpc.rs from multisig.rs
Move RPC helpers, Keypair struct, and Solana CLI config functions into a
dedicated rpc module shared by both multisig.rs and deploy.rs. Add
Keypair::generate() for programmatic keypair creation.
